Delay relation-joined until unit has address
