+++ apiserver/facades/agent/uniter/uniter.go
@@ -1260,6 +1260,17 @@ func (u *UniterAPI) Refresh(args params.Entities) (params.UnitRefreshResults, er
 		if canRead(tag) {
 			var unit *state.Unit
 			if unit, err = u.getUnit(tag); err == nil {
+				// IAAS units always have an address.
+				result.Results[i].HasAddress = u.m.Type() == state.ModelTypeIAAS
+				// For CAAS models we need to check.
+				if !result.Results[i].HasAddress {
+					addresses, err := unit.AllAddresses()
+					if err != nil {
+						result.Results[i].Error = common.ServerError(err)
+						continue
+					}
+					result.Results[i].HasAddress = len(addresses) > 0
+				}
 				result.Results[i].Series = unit.Series()
 				result.Results[i].Life = params.Life(unit.Life().String())
 				result.Results[i].Resolved = params.ResolvedMode(unit.Resolved())
